refactor: use loading context in all service calls
