+++ app/src/pages/FindUsersPages/styles.js
@@ -6,10 +6,6 @@ const styles = StyleSheet.create({
         flex: 1,
         padding: 20,
     },
-    loadingContainer: {
-        height: '100%',
-        justifyContent: 'center',
-    },
 
     profileImage: {
         width: 100,
@@ -18,10 +14,6 @@ const styles = StyleSheet.create({
         borderRadius: 50,
         marginHorizontal: 10,
         alignSelf: 'center',
-    },
-
-    loading: {
-        marginVertical: 20,
     },
 
     card: {
