+++ app/src/pages/HelpPages/CreateHelpRequest/styles.js
@@ -42,13 +42,6 @@ const styles = StyleSheet.create({
         color: colors.primary,
     },
 
-    loading: {
-        flex: 1,
-        height: '100%',
-        justifyContent: 'center',
-        flexDirection: 'row',
-        padding: 10,
-    },
     descriptionInput: {
         marginTop: 20,
     },
