+++ app/src/pages/Profile/EditProfileFields/CEPField/styles.js
@@ -44,11 +44,6 @@ export default StyleSheet.create({
         flexGrow: 1,
         justifyContent: 'center',
     },
-    scrollLoading: {
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'center',
-    },
     viewMargin: {
         marginVertical: 6,
     },
