+++ app/src/store/contexts/loadingContext.jsx
@@ -4,7 +4,7 @@ import { LoadingIndicator } from '../../components/LoadingIndicator';
 export const LoadingContext = createContext({});
 
 export const LoadingContextProvider = ({ children }) => {
-    const [isLoading, setIsLoading] = useState(true);
+    const [isLoading, setIsLoading] = useState(false);
     return (
         <LoadingContext.Provider value={{ isLoading, setIsLoading }}>
             {isLoading && <LoadingIndicator />}
